+++ eslint.config.js
@@ -2,8 +2,8 @@ import node from './packages/eslint-config-node/index.mjs'
 import optional from './packages/eslint-config-node/optional.mjs'
 import style from './packages/eslint-config-node/style.mjs'
 
-/** @type {Array<import("eslint").Linter.FlatConfig>} */
-const configs = [{
+/** @type {Array<import("eslint").Linter.Config>} */
+const config = [{
   linterOptions: {
     reportUnusedDisableDirectives: true,
   },
@@ -11,15 +11,15 @@ const configs = [{
     'node_modules',
     '!.*.js',
   ],
+  settings: {
+    'import/resolver': { typescript: {} },
+  },
 },
 node,
 optional,
 style,
-{ files: ['**/*.js', '**/*.mjs'], languageOptions: { sourceType: 'module' } },
-{ files: ['**/*.cjs'], languageOptions: { sourceType: 'commonjs' } },
-{ rules: {
-  // TODO: Figure out why ESLint cannot resolve the @typescript-eslint packages
-  'import/no-unresolved': 'off',
-} }]
+{
+  rules: {},
+}]
 
-export default configs
+export default config