+++ packages/eslint-config-base/style.mjs
@@ -1,7 +1,7 @@
 import imports from 'eslint-plugin-import'
 import stylistic from '@stylistic/eslint-plugin'
 
-/** @type {import("eslint").Linter.FlatConfig} */
+/** @type {import("eslint").Linter.Config} */
 const config = {
   plugins: {
     import: imports,
@@ -166,6 +166,8 @@ const config = {
         'R',
         // jQuery 🤮
         '$',
+        // vitest in-test context: `it('works', t => { t.expect(1).toBe(1) })`
+        't',
       ],
     }],
 
@@ -227,7 +229,7 @@ const config = {
 
     // Enforce a particular style for multiline comments
     // Prefer consecutive line comments for multiline comment block
-    'multiline-comment-style': ['warn', 'separate-lines'],
+    '@stylistic/multiline-comment-style': ['warn', 'separate-lines'],
 
     // Enforce newlines between operands of ternary expressions
     // Enforce either the whole ternary on a single line or each part of the ternary on its own line