+++ packages/eslint-config-typescript/index.mjs
@@ -2,7 +2,7 @@ import base from '@strv/eslint-config-base'
 import ts from '@typescript-eslint/eslint-plugin'
 import parser from '@typescript-eslint/parser'
 
-/** @type {import("eslint").Linter.FlatConfig} */
+/** @type {import("eslint").Linter.Config} */
 const config = {
   plugins: {
     ...base.plugins,
@@ -10,6 +10,9 @@ const config = {
   },
   languageOptions: {
     parser,
+    parserOptions: {
+      projectService: true,
+    },
   },
   settings: {
     'import/resolver': { typescript: {} },
@@ -247,7 +250,7 @@ const config = {
     // the `Error` object as base objects for user-defined exceptions. The fundamental benefit of
     // `Error` objects is that they automatically keep track of where they were built and
     // originated.
-    '@typescript-eslint/no-throw-literal': [base.rules['no-throw-literal'], {
+    '@typescript-eslint/only-throw-error': [base.rules['no-throw-literal'], {
       allowThrowingAny: false,
       allowThrowingUnknown: false,
     }],