+++ packages/eslint-config-typescript/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [6.0.0-alpha.2](https://github.com/strvcom/code-quality-tools/compare/@strv/eslint-config-typescript@6.0.0-alpha.1...@strv/eslint-config-typescript@6.0.0-alpha.2) (2025-01-24)
+
+
+### Features
+
+* enable typescript project service 👏 ([72b29bf](https://github.com/strvcom/code-quality-tools/commit/72b29bff9a1473d49fd644b5cc93fae0905e4353))
+
+
+
+
+
+# [6.0.0-alpha.1](https://github.com/strvcom/code-quality-tools/compare/@strv/eslint-config-typescript@6.0.0-alpha.0...@strv/eslint-config-typescript@6.0.0-alpha.1) (2025-01-24)
+
+
+### Bug Fixes
+
+* rename `@typescript-eslint/no-throw-literal` -> `only-throw-error` ([c67389c](https://github.com/strvcom/code-quality-tools/commit/c67389c05befb058c89755126802ff4faec4b612))
+
+
+
+
+
+# [6.0.0-alpha.0](https://github.com/strvcom/code-quality-tools/compare/@strv/eslint-config-typescript@5.0.0...@strv/eslint-config-typescript@6.0.0-alpha.0) (2025-01-24)
+
+
+### Features
+
+* move to ESLint v9 👏 ([4de1c1f](https://github.com/strvcom/code-quality-tools/commit/4de1c1f7543898cddc4178afc7222e369b6ea018))
+
+
+### BREAKING CHANGES
+
+* Projects must now use ESLint v9 in order to use these rulesets.
+
+
+
+
+
 # [5.0.0](https://github.com/strvcom/code-quality-tools/compare/@strv/eslint-config-typescript@4.1.2...@strv/eslint-config-typescript@5.0.0) (2024-02-16)